Uploaded files
'.hx' is a temporary file extension as short for "High level code of unnamed language 'X'."
The syntax is also available. (Not exactly understandable though. Will probably work on that too.)
The 'Alpha1' was for reading the machine code, since I can't afford to make a CPU with a custom architecture.
'V6' is the main code that reads the .hx file and converts it into an 'lx' file.